Updated Trend Function
Updated trend function so that it creates a common weight variable and cluster variable for each school that differ's based on the district and variables information each year. Test Code was created to be able to run the function on each district to make sure the function works. Added the HS Snapshot trend variable to git repo since the function needs it to create the table.