Up Dated Codes so that the function can work with the state data.
For the trend function, I added more logic to incorporate the strata for the state that isn't just the year variable. Additionally, incorporated the use of fpc. In the trend_report I renamed the report and deleted the old one. I updated the data reading portion so that it handles the states data too. In the report automation code, I added code so that the state report would be rendered as well as read in the state data sets.